requirements.txt
opencv-python==4.8.1.78
deepface==0.0.79
pyttsx3==2.90
Pillow==10.1.0
tk==0.1.0
tensorflow==2.15.0
flask==3.0.0
numpy==1.24.3
mysql-connector-python==8.2.0
google-generativeai==0.3.2
beautifulsoup4==4.12.2
requests==2.31.0
gunicorn==21.2.0
